bicycle-cycle-local: Don't also hide top-level when hiding a block
We do need to also call `outline-show-entry' after `hs-show-block'
for the reason given in 42a5db3514019d539500a67f913411f5533a1eb3,
but that does not mean that we should also call `outline-hide-entry'
after `hs-hide-block'.

The `outline-show-entry' is only relevant for top-level code blocks,
but does no harm for lower level code blocks because the respective
top-level block ought to be expanded anyway.

The same does not apply to `outline-hide-entry', because in this case
the top-level block is expanded and so hiding does have an effect.

Closes #5.
Closes #6.