Improve description of global cycling states
Number the steps the same as the equivalent local steps, which also
means we have to mention the second step, which isn't implemented for
global cycling.  I think this step would only very seldom be useful,
and much more often would lead to users thinking they have reached
the TOC (third) state, when in fact they have not.