Wrap calls to hs-{hide,show}-block with save-excursion
In the past these functions moved point forward exactly one character,
which is undesirable and probably a bug.  We countered that by moving
backward exactly one character.

Now these functions move forward to the end of the hidden area, which
is also undesirable and more destructive, and most likely also not
intended.

Deal with both variations of undesirable movement by wrapping the calls
to these function with `save-excursion'.